requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
reportlab==4.0.4
Pillow==11.2.1
docx2pdf==0.1.8
gunicorn==21.2.0
requests==2.31.0 